+++ modules/supabase.mjs
@@ -5,9 +5,10 @@ export const name = "makeup-api-supabase";
 // TODO(chantel): fill in the blanks
 // Create a single supabase client for interacting with your database
 const supabase = createClient(
-  "https://xyzcompany.supabase.co",
-  "public-anon-key"
-);
+    "https://axlkpyetnmbcnzncpwvp.supabase.co",
+  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImF4bGtweWV0bm1iY256bmNwd3ZwIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MzQwNjAzMDgsImV4cCI6MjA0OTYzNjMwOH0.hKKyYuNkQ7NyRJif0sZZfps_5Fr8T-4LgYN6s-UGHtY"
+)
+export default supabase;
 
 /*
 Chris' thoughts on the tables:
@@ -53,15 +54,39 @@ function Palette(name, data) {
  * HINT: this one will need at least 2 queries
  * @return the associated user_id
  */
-export function addUser(name, email, age) {
-  return 1;
+export async function addUser(name, email, age) {
+    const { data: existingUser, error: selectError } = await supabase
+    .from("users")
+    .select("user_id")
+    .eq("email", email);
+
+  if (selectError) throw selectError;
+
+  if (existingUser.length > 0) {
+    return existingUser[0].user_id; 
+  }
+
+  const { data: newUser, error: insertError } = await supabase
+    .from("users")
+    .insert({ name, email, age })
+    .select("user_id");
+
+  if (insertError) throw insertError;
+
+  return newUser[0].user_id; 
 }
+
 
 /**
  * @return user data based on user_id
  */
-export function getUser(user_id) {
-  // TODO(chantel): replace with actual query
+export async function getUser(user_id) {
+    const { data, error } = await supabase
+    .from("users")
+    .select("name, email, age")
+    .eq("user_id", user_id)
+    .single(); 
+  if (error) throw error;
   return new User("Chris Johnson", "test@example.com", 22);
 }
 
@@ -70,23 +95,51 @@ export function getUser(user_id) {
  * user_palettes table
  * @returns the associated palette_id
  */
-export function addPalette(user_id, palette_name, palette_data) {
-  return 1;
+export async function addPalette(user_id, palette_name, palette_data) {
+    const { data: palette, error: insertPaletteError } = await supabase
+    .from("palettes")
+    .insert({ palette_name, palette_data })
+    .select("palette_id");
+
+  if (insertPaletteError) throw insertPaletteError;
+
+  const palette_id = palette[0].palette_id;
+
+  const { error: linkError } = await supabase
+    .from("user_palettes")
+    .insert({ user_id, palette_id });
+
+  if (linkError) throw linkError;
+
+  return palette_id;
 }
 
 /**
  * @return palette based on palette_id
  */
 export function getPalette(palette_id) {
-  return new Palette("Foo", 113, 492, 1004);
-}
+    const { data, error } = supabase
+    .from("palettes")
+    .select("palette_name, palette_data")
+    .eq("palette_id", palette_id)
+    .single();
 
+  if (error) throw error;
+
+  return new Palette(data.palette_name, data.palette_data);
+}
 /**
  * @return all palettes related to a user_id (array)
  */
 export function getPalettesByUser(user_id) {
-  return [
-    new Palette("Foobar?", [1, 2, 3, 4, 5]),
-    new Palette("FizzBuzz", [1, 2, 4, 6, 7]),
-  ];
+    const { data, error } = supabase
+    .from("user_palettes")
+    .select(`
+      palettes (palette_name, palette_data)
+    `)
+    .eq("user_id", user_id);
+
+  if (error) throw error;
+
+  return data.map((entry) => new Palette(entry.palettes.palette_name, entry.palettes.palette_data));
 }
